+++ app/controllers/design_options_controller.rb
@@ -201,7 +201,7 @@ class DesignOptionsController < ApplicationController
 
     params.require(:domain).permit(
       :name, :display_name, :description, :user_id,
-      option_tokens: [:name, :value, :description, :missing_code, :site_id, :domain_option_id]
+      option_tokens: [:name, :value, :description, :missing_code, :site_id, :domain_option_id, :archived]
     )
   end
 end